Added a Scale option to StepsView. Changed OptionsToAssociativeList so that numbers and constants are not automatically converted into strings. Version bump.

+++ src/library_apps/mathpiper4/src/org/mathpiper/builtin/functions/optional/TreeView.java
@@ -187,7 +187,7 @@ public class TreeView extends BuiltinFunction {
         
 
         
-        int viewScale = (int) ((Double)userOptions.get("Scale")).doubleValue();
+        double viewScale = ((Double)userOptions.get("Scale")).doubleValue();
         
         JPanel panel = new JPanel();
         panel.setLayout(new BorderLayout());
